# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: EDP Control Plane API
  description: |
    Enterprise data control plane for registering sources and datasets,
    authoring and deploying pipelines, enforcing data quality and governance
    policy, tracking lineage and impact, orchestrating runs, and recording an
    append-only audit trail.

    ## Conventions

    * All collection endpoints are paginated with `limit` and `offset`, and
      sorted with `sort_by` and `sort_dir`. See the shared parameters.
    * Errors always use the same envelope: `error.code`, `error.message`,
      `error.details` and a top-level `request_id` that matches the
      `X-Request-Id` response header.
    * Mutations are `PATCH` rather than `PUT`, so a caller updates only the
      fields it means to change.
    * Request bodies are validated before any state is written; a validation
      failure returns `422` with one entry per offending field.
  version: 0.1.0
  contact:
    name: Data Platform Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://edp.internal/api/v1
    description: Internal deployment
security:
  - bearerAuth: []
...
paths:
...
  /pipeline-runs:
    get:
      tags: [execution]
      operationId: listPipelineRuns
      summary: List pipeline runs
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: pipeline_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          description: Repeatable.
          schema: { type: array, items: { type: string, enum: [queued, running, success, failed, cancelled, skipped] } }
        - name: active_only
          in: query
          description: Only queued or running runs.
          schema: { type: boolean }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: A page of runs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PipelineRunPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [execution]
      operationId: triggerPipelineRun
      summary: Trigger a pipeline run
      description: |
        Requires the pipeline to be deployed. Returns `409` when the pipeline's
        `max_active_runs` is already reached, so a caller can back off rather than
        queue without bound.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pipeline_id]
              properties:
                pipeline_id: { type: string, format: uuid }
                triggered_by: { type: string }
                config:
                  type: object
                  additionalProperties: true
                note: { type: string }
      responses:
        '201':
          description: The created run, in `queued` state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PipelineRun' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503':
          description: No orchestrator is configured.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /pipeline-runs/by-key/{runID}:
    parameters:
      - name: runID
        in: path
        required: true
        description: The orchestrator run key, not the internal UUID.
        schema: { type: string }
    get:
      tags: [execution]
      operationId: getPipelineRunByKey
      summary: Fetch a run by orchestrator key
      responses:
        '200':
          description: The run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PipelineRun' }
        '404': { $ref: '#/components/responses/NotFound' }
  /pipeline-runs/{runID}:
    parameters:
      - name: runID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [execution]
      operationId: getPipelineRun
      summary: Fetch a run
      responses:
        '200':
          description: The run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PipelineRun' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /pipeline-runs/{runID}/tasks:
    parameters:
      - name: runID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [execution]
      operationId: listTaskRuns
      summary: List task runs for a pipeline run
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema: { type: array, items: { type: string } }
      responses:
        '200':
          description: A page of task runs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskRunPage' }
        '404': { $ref: '#/components/responses/NotFound' }
  /pipeline-runs/{runID}/tasks/{taskRunID}:
    parameters:
      - name: runID
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: taskRunID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [execution]
      operationId: getTaskRun
      summary: Fetch a task run
      responses:
        '200':
          description: The task run, including log references.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskRun' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orchestrator/runs/{runID}/sync:
    parameters:
      - name: runID
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [execution]
      operationId: syncOrchestratorRun
      summary: Reconcile one run from the orchestrator
      description: |
        The callback endpoint Airflow calls when a run finishes. Idempotent, so a
        retried callback does not corrupt the recorded state. An unknown run key
        returns `404` so the orchestrator stops retrying.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state: { type: string }
                start_date: { type: string, format: date-time }
                end_date: { type: string, format: date-time }
                message: { type: string }
      responses:
        '200':
          description: The updated run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PipelineRun' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orchestrator/tasks/sync:
    post:
      tags: [execution]
      operationId: syncOrchestratorTask
      summary: Reconcile one task run from the orchestrator
      description: Idempotent, in the same way as the run callback.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_key, state]
              properties:
                run_id: { type: string }
                task_key: { type: string }
                try_number: { type: integer, minimum: 1 }
                state: { type: string }
                start_date: { type: string, format: date-time }
                end_date: { type: string, format: date-time }
                log_url: { type: string }
      responses:
        '200':
          description: The updated task run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskRun' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT verified against the configured JWKS. Issuer, audience and expiry are
        all checked. When authentication is disabled the API runs without any
        middleware and every request is attributed to a development actor.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Rows to skip.
      schema: { type: integer, minimum: 0, default: 0 }
    SortBy:
      name: sort_by
      in: query
      description: Column to sort on. Unknown columns are rejected.
      schema: { type: string }
    SortDir:
      name: sort_dir
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
...
  responses:
    BadRequest:
      description: The query string or path parameters were malformed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The request conflicts with current state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: The body failed validation. Every offending field is listed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - CONFLICT
                - VALIDATION_FAILED
                - RATE_LIMITED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - NOT_IMPLEMENTED
            message:
              type: string
              description: Human-readable summary. Safe to show a user.
            details:
              type: array
              description: Present for validation failures.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
        request_id:
          type: string
          format: uuid
          description: Matches the X-Request-Id header, for correlating with logs.
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
...
    PipelineRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        pipeline_id: { type: string, format: uuid }
        run_id: { type: string, description: The orchestrator run key. }
        dag_id: { type: string }
        status: { type: string, enum: [queued, running, success, failed, cancelled, skipped] }
        trigger_type: { type: string, enum: [manual, scheduled, api, upstream] }
        config: { type: object, additionalProperties: true }
        started_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, nullable: true }
        quality_gate_status: { type: string, nullable: true }
        quality_gate_rule_id: { type: string, format: uuid, nullable: true }
        triggered_by: { type: string }
        created_at: { type: string, format: date-time }
    TaskRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        task_key: { type: string }
        try_number: { type: integer }
        status: { type: string, enum: [queued, running, success, failed, skipped, upstream_failed] }
        started_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, nullable: true }
        log_url: { type: string }
        message: { type: string }
...
    PipelineRunPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PipelineRun' } }
    TaskRunPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/TaskRun' } }
...
    Page:
      type: object
      description: Every list endpoint returns the same envelope.
      required: [items, meta]
      properties:
        items:
          type: array
          items: {}
        meta: { $ref: '#/components/schemas/PageMeta' }